test: share tracing capture helper
